Make sure all integral type members implement v() method

+++ src/main/java/nom/bdezonia/zorbage/type/data/int1/UnsignedInt1Member.java
@@ -85,7 +85,7 @@ public final class UnsignedInt1Member
 		get(value);
 	}
 
-	public int v() {
+	public byte v() {
 		return v;
 	}
